+++ views/pages/listStacks.ejs
@@ -23,7 +23,7 @@
         <row>
             <div class="col-md-12">
                 <h1>Stacks</h1>
-                <p>Manage Stacks</p>
+                <p>View Stacks</p>
             </div>
         </row>
 
@@ -34,7 +34,7 @@
     <div class="col-md-12">
         <ol class="breadcrumb" style="margin-top: 10px;">
             <li><a href="/">Home</a></li>
-            <li class="active">Manage stacks</li>
+            <li class="active">View stacks</li>
         </ol>
     </div>
